refactor fixture component
