+++ app/src/main/java/com/diegopizzo/whosplaying/ui/component/attr/Padding.kt
@@ -6,6 +6,6 @@ import androidx.compose.ui.unit.dp
 val defaultPadding: Dp = 16.dp
 val tinyPadding: Dp = 4.dp
 val smallPadding: Dp = 8.dp
-val mediumPadding: Dp = 32.dp
+val mediumPadding: Dp = 24.dp
 val largePadding: Dp = 32.dp
