+++ app/src/main/java/com/diegopizzo/whosplaying/ui/component/common/CustomView.kt
@@ -2,10 +2,15 @@ package com.diegopizzo.whosplaying.ui.component.common
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
+import androidx.compose.material.Divider
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
 import coil.compose.rememberImagePainter
 import com.diegopizzo.whosplaying.R
+import com.diegopizzo.whosplaying.ui.component.attr.teal700
 
 @Composable
 fun ComposeImage(
@@ -23,4 +28,9 @@ fun ComposeImage(
         contentDescription = null,
         modifier = modifier
     )
+}
+
+@Composable
+fun MyDivider(modifier: Modifier = Modifier, color: Color = teal700, thickness: Dp = 1.dp) {
+    Divider(color = color, modifier = modifier, thickness = thickness)
 }
