+++ app/src/main/java/com/diegopizzo/whosplaying/ui/component/common/Text.kt
@@ -45,8 +45,14 @@ fun MediumText(text: String, modifier: Modifier = Modifier) {
 }
 
 @Composable
-fun DefaultText(text: String, modifier: Modifier = Modifier) {
-    Text(text = text, fontSize = 22.sp, color = MaterialTheme.colors.textColor, modifier = modifier)
+fun DefaultText(text: String, modifier: Modifier = Modifier, textAlign: TextAlign? = null) {
+    Text(
+        text = text,
+        fontSize = 22.sp,
+        color = MaterialTheme.colors.textColor,
+        modifier = modifier,
+        textAlign = textAlign
+    )
 }
 
 @Composable
