+++ app/src/main/java/com/diegopizzo/whosplaying/ui/detailsscreen/DetailsScreenView.kt
@@ -10,8 +10,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.diegopizzo.network.model.*
 import com.diegopizzo.whosplaying.R
-import com.diegopizzo.whosplaying.ui.component.ComposeEvent
-import com.diegopizzo.whosplaying.ui.component.ComposeFixturesDetails
+import com.diegopizzo.whosplaying.ui.component.fixture.ComposeEvent
+import com.diegopizzo.whosplaying.ui.component.fixture.ComposeFixturesDetails
 import com.diegopizzo.whosplaying.ui.component.attr.smallPadding
 import com.diegopizzo.whosplaying.ui.component.common.MediumText
 import com.diegopizzo.whosplaying.ui.component.common.MyCard
