+++ app/src/main/res/layout/item_fixture.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.diegopizzo.whosplaying.ui.component.FixtureComponent xmlns:android="http://schemas.android.com/apk/res/android"
+<com.diegopizzo.whosplaying.ui.component.fixture.FixtureComponent xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/item_fixture_id"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
